+++ demo/restoration_video_demo.py
@@ -8,8 +8,7 @@ import numpy as np
 import torch
 
 from mmedit.apis import init_model, restoration_video_inference
-from mmedit.engine import tensor2img
-from mmedit.utils import modify_args
+from mmedit.utils import modify_args, tensor2img
 
 VIDEO_EXTENSIONS = ('.mp4', '.mov')
