+++ mmedit/engine/__init__.py
@@ -1,7 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .hooks import *  # noqa: F401, F403
 from .logging import *  # noqa: F401, F403
-from .misc import *  # noqa: F401, F403
 from .optimizers import *  # noqa: F401, F403
 from .runner import *  # noqa: F401, F403
 from .schedulers import *  # noqa: F401, F403
