+++ mmedit/models/editors/flavr/flavr.py
@@ -1,7 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from mmedit.engine.misc import tensor2img
 from mmedit.models.base_models import BasicInterpolator
 from mmedit.registry import MODELS
+from mmedit.utils import tensor2img
 
 # TODO tensor2img will be move
