+++ mmedit/utils/__init__.py
@@ -5,10 +5,11 @@ from .io_utils import MMGEN_CACHE_DIR, download_from_url
 from .logger import print_colored_log
 from .sampler import get_sampler
 from .setup_env import register_all_modules
+from .tensor2img import tensor2img
 from .typing import ForwardInputs, LabelVar, NoiseVar, SampleList
 
 __all__ = [
     'modify_args', 'print_colored_log', 'register_all_modules',
     'ForwardInputs', 'SampleList', 'NoiseVar', 'LabelVar', 'MMGEN_CACHE_DIR',
-    'download_from_url', 'get_sampler'
+    'download_from_url', 'get_sampler', 'tensor2img'
 ]
